test(prio-nav): add failing tests for issue with automatic priority menu
Needed to add the flex-shrink: 0; since otherwise the items would just
shrink to fit instead of actually overflowing as the tests wanted.

Also one of the tests needs to get a bit more size, 100px is not enough
to show an item and the priority menu, so I changed it to 150px.